refactor(backend): 按分ロジックを services/settlement/ に集約（#105-6）
itemSplitAllocation / itemSplitService を settlement 配下へ移動し、receiptController のエントリポイントを更新。旧 receiptSplitService は re-export のみ残す。
